// src/types/product.ts
export type ProductToCreate = {
  name: string;
  description?: string;
  upc_code?: string;
  gtin?: string;
  reference_gtin?: string;
  visit_type?:
    | "cbtSleep"
    | "notApplicable"
    | "metabolomics" 
    | "tinnitus"
    | "gym"
    | "exerciseDiet"
    | "orthopedic"
    | "alcohol"
    | "airPurification"
    | "vaginalHealth"
    | "menstrual"
    | "canopy"
    | "alopecia"
    | "genate"
    | "weightBlanket"
    | "bedJet"
    | "siderAl"
    | "sunwinkPowder"
    | "wavWatch"
    | "bodyComplete"
    | "ageRate"
    | "nitrousOxide"
    | "happyV"
    | "groupChat"
    | "icalmAnxiety"
    | "redBloom"
    | "foodom"
    | "branchErgonomicFurniture"
    | "curalife"
    | "eloraInfantWellness"
    | "nutriHealth"
    | "olipop"
    | "goldIntimate"
    | "touchStoneEssentials"
    | "utiva"
    | "sleepGeekz"
    | "figBrew"
    | "auBabyBlanket"
    | "oshWellness"
    | "currentBodyRedLight"
    | "mito"
    | "circularRing"
    | "lymaRedLight"
    | "goodAirRx"
    | "tastermonial"
    | "karunaHome"
    | "ergoStandingChair"
    | "jbaGlucoseControl"
    | "bloomNutrition"
    | "buoyDrops"
    | "luxeWonderWig"
    | "saunaMarketplace"
    | "amrioreEyewear"
    | "pivotOrthoShoe"
    | "lumenCynergy"
    | "roga"
    | "pulsetto"
    | "mitoRedLight"
    | "gutPersonal"
    | "goFlaus"
    | "myHixel"
    | "calmigo"
    | "dotFit"
    | "stripesBeauty"
    | "mixHers"
    | "pmd"
    | "positivityWithPurpose"
    | "techRing"
    | "popVeneers"
    | "vertaClean"
    | "lumen"
    | "coolingBed"
    | "posture"
    | "supplements"
    | "sleep"
    | "redLightTherapy"
    | "fitness"
    | "smartRing"
    | "womensVaginalHealth"
    | "fertilitySupport"
    | "femaleReproduction"
    | "pregnancyLiterature"
    | "iceBath"
    | "orthopedicShoes"
    | "sexualHealth"
    | "glucose"
    | "metabolicTest"
    | "skinCare"
    | "oralHealth"
    | "blueLightGlasses"
    | "anxiety"
    | "brainHealth"
    | "babyMonitor"
    | "compressionSocks"
    | "waterPurification"
    | "medSpa"
    | "essentialOils"
    | "sleepBuds"
    | "latchLight"
    | "nutritionist"
    | "rairflow"
    | "enduranceTraining"
    | "hydration"
    | "hairGrowth"
    | "eD"
    | "postureFitness"
    | "childDevelopment"
    | "adaptiveClothing"
    | "sleepConsulting"
    | "hairRemoval"
    | "menopause"
    | "maleFertility"
    | "anxietyHealth"
    | "bidets"
    | "speechHealth";
  url?: string;
  hsa_fsa_eligibility?: "not_eligible" | "auto_substantiation" | "private_label" | "letter_of_medical_necessity" | "prescription" | "vision" | "service";
  metadata?: Record<string, unknown>;
  // Our e-commerce additions for display (not sent to Flex)
  price: number;
  image: string;
};
// ...
export type CartItem = ProductToCreate & {
  quantity: number;
};